domain-core/rdata: added Ord + Eq to RtypeRecordData
The RDATA of the same type need to be ordered for [RFC4034, 6.4](https://tools.ietf.org/html/rfc4034#section-6.3)
purposes. The canonical ordering is the same as bytewise comparison as long as all fields in the structure appear
in the same order as in the wire format. The exception is Serial which implements a different Ord semantics,
so I changed it to use it's underlying size in struct fields, and convert to Serial in accessor methods.